// src/interfaces/isencao.ts
export interface Isencao_class {
  codigo: number;
  nome: string;
  cpf: string;
  email: string;
  rg: string;
  telefone1: string;
  telefone2: string;
  iptu: string;
  cad_oficio: boolean;
  anotacoes: string;
  dt_casdastro: string;
  nome_conjuge: string;
  ident_conjuge: string;
  data_nascimento?: null;
  ident_requisitante: string;
  nome_requisitante: string;
  endereco: string;
  numero: number;
  complemento: string;
  bairro: string;
  municipio: string;
  cep: string;
  uf: string;
  situacao: boolean;
  dt_gravacao: string;
  anexo: string;
  uc: string;
  anexos_sts: string;
  anexos_status: number;
}